+++ WEB-INF/classes/SignupPage.java
@@ -5,14 +5,13 @@
 // 必要なパッケージの指定
 import java.io.*;
 import javax.servlet.*;
-import javax.servlet.annotation.MultipartConfig;
-import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.*;
+import javax.servlet.annotation.MultipartConfig;
 import java.sql.*;
 
 @MultipartConfig(location="", maxFileSize=1048576)
 public class SignupPage extends HttpServlet {
-	public void doGet(HttpServletRequest hreq,	// リクエスト
+	public void doPost(HttpServletRequest hreq,	// リクエスト
 			  HttpServletResponse hres)	// レスポンス
 		throws ServletException, IOException {
 		// リクエストパラメータの文字エンコーディング指定
@@ -31,12 +30,13 @@ public class SignupPage extends HttpServlet {
 		String mail = (String)session.getAttribute("mail");
 		String twitter = (String)session.getAttribute("twitter");
 		String facebook = (String)session.getAttribute("facebook");
-		Part iconUrl = (Part)session.getAttribute("icon_url");
+		String fileName = (String)session.getAttribute("fileName");
 
 		// HTMLテキストの出力
 		out.println("<html><head><meta http-equiv=\"Pragma\" content=\"no-cache\">");
 		out.println("<meta http-equiv=\"Expires\" content=\"-1\">");
 		out.println("<meta http-equiv=\"Content-Type\" content=\"text/html; charset=utf-8\">");
+		out.println("</head><body>");
 		// ここまでは以降もそのまま使う
 		Connection db = null;
 		String db_name = "circle_triangle";
@@ -57,14 +57,11 @@ public class SignupPage extends HttpServlet {
 				out.println("</body></html>");
 			} else {
 				// SQL 文を query に格納
-				query = "insert into users set id='" + userId + "', university_id=" + universityId + ", sex=" + sex + ",password='" + pass + "',mail='" + mail + "',twitter='" + twitter + "',facebook='" + facebook + "'";//, icon_url='" + iconUrl+"'";
+				query = "insert into users set id='" + userId + "', university_id=" + universityId + ", sex=" + sex + ",password='" + pass + "',mail='" + mail + "',twitter='" + twitter + "',facebook='" + facebook + "', icon_url='" + fileName + "'";
 				// SQL 文を実行し挿入した数が返る
 				int num = st.executeUpdate(query);
 				if(num > 0) {
 					//登録成功
-					//画像保存処理
-					String name = this.getFileName(iconUrl);
-					iconUrl.write(getServletContext().getRealPath("/uploads") + "/users/" + name);
 					session.setAttribute("Login",1);
 					session.setAttribute("IdData",userId);
 					out.println("データが登録されました．");
@@ -87,17 +84,5 @@ public class SignupPage extends HttpServlet {
 			} catch(Exception e){}
 		}
 	out.println("</body></html>");
-	}
-
-	private String getFileName(Part part) {
-		String name = null;
-		for (String dispotion : part.getHeader("Content-Disposition").split(";")) {
-			if (dispotion.trim().startsWith("filename")) {
-				name = dispotion.substring(dispotion.indexOf("=") + 1).replace("\"", "").trim();
-				name = name.substring(name.lastIndexOf("\\") + 1);
-				break;
-			}
-		}
-		return name;
 	}
 }
